rename installables/ -> mods/
