+++ src/extensionConsole/pages/blueprints/BlueprintsPage.tsx
@@ -17,7 +17,7 @@
 
 import React, { useEffect } from "react";
 import BlueprintsPageLayout from "@/extensionConsole/pages/blueprints/BlueprintsPageLayout";
-import useInstallables from "@/installables/useInstallables";
+import useInstallables from "@/mods/useInstallables";
 import { useTitle } from "@/hooks/title";
 import { ErrorDisplay } from "@/layout/ErrorDisplay";
 import { reportEvent } from "@/telemetry/events";
