+++ src/extensionConsole/pages/blueprints/BlueprintsPageLayout.test.tsx
@@ -18,7 +18,7 @@
 import React from "react";
 import { render } from "@/extensionConsole/testHelpers";
 import BlueprintsPageLayout from "@/extensionConsole/pages/blueprints/BlueprintsPageLayout";
-import { type Installable } from "@/installables/installableTypes";
+import { type Installable } from "@/mods/installableTypes";
 import { waitForEffect } from "@/testUtils/testHelpers";
 import { act, screen } from "@testing-library/react";
 import blueprintsSlice from "@/extensionConsole/pages/blueprints/blueprintsSlice";
