+++ src/extensionConsole/pages/blueprints/BlueprintsPageSidebar.tsx
@@ -18,7 +18,7 @@ import {
   faUsers,
 } from "@fortawesome/free-solid-svg-icons";
 import { type TableInstance } from "react-table";
-import { type InstallableViewItem } from "@/installables/installableTypes";
+import { type InstallableViewItem } from "@/mods/installableTypes";
 import useFlags from "@/hooks/useFlags";
 import { appApi, useGetStarterBlueprintsQuery } from "@/services/api";
 import { kebabCase } from "lodash";
