+++ src/extensionConsole/pages/blueprints/hooks/useActivateAction.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type InstallableViewItem } from "@/installables/installableTypes";
+import { type InstallableViewItem } from "@/mods/installableTypes";
 import { useDispatch } from "react-redux";
 import { isExtension } from "@/utils/installableUtils";
 import { reportEvent } from "@/telemetry/events";
