+++ src/extensionConsole/pages/blueprints/hooks/useReactivateAction.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type InstallableViewItem } from "@/installables/installableTypes";
+import { type InstallableViewItem } from "@/mods/installableTypes";
 import { useDispatch } from "react-redux";
 import useFlags from "@/hooks/useFlags";
 import { isExtension, isExtensionFromRecipe } from "@/utils/installableUtils";
