+++ src/mods/InstallableIcon.module.scss
