+++ src/mods/hooks/useInstallablePermissions.ts
