+++ src/mods/hooks/useMarketplaceUrl.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type InstallableViewItem } from "@/installables/installableTypes";
+import { type InstallableViewItem } from "@/mods/installableTypes";
 import { MARKETPLACE_URL } from "@/utils/strings";
 
 function useMarketplaceUrl(
